requirements.txt
requests
pyyaml
beautifulsoup4
lxml
